Fix Render Graph Not Resizing With Window
This change modifies the engine code to make sure the render graph is
rebuilt whenever the window size changes since it will be dependent on
it in most cases. The viewport/scissor code for the copy shader was also
updated inside the renderer in order to make this work.

This change also cleans up some vk warnings that were showing up in
renderdoc related to gfx descriptor set bindings.