Audio Track Functionality
This commit enables support for playback of an audio track in the
engine code. Playback can be controlled with the spacebar (play/pause)
and the left/right arrow keys (seek) with shift + alt as speed
modifiers.

This commit also includes a small test audio file so the playback
functionality can be tested.