Expose Time to Shaders
This change exposes the current track time to shaders. The "Yellow"
shader has been updated to use time to manipulate the test colors.

This change also fixes a Vulkan issue where the main descriptor set was
being updated after it was bound to a command buffer. The fix for this
was simply making sure the render graph image view descriptor slot gets
updated before any rendering is performed on the main command buffer.