Always Enable Vulkan Debug Message When Available
This change modifies the vk device code to ensure that the debug message
extension is always loaded when it's available. Even when the validation
layers are not enabled.

This will allow release builds of the engine to have proper gpu event
labels in debug/profiling tools and will also aid in the debug process
for deployed executables.