Cleanup from @Party (#35)
* Always Enable Vulkan Debug Message When Available

This change modifies the vk device code to ensure that the debug message
extension is always loaded when it's available. Even when the validation
layers are not enabled.

This will allow release builds of the engine to have proper gpu event
labels in debug/profiling tools and will also aid in the debug process
for deployed executables.

* Release Build Updates

This change modifies the engine to only allow keyboard based audio
control in debug builds, and to automatically start the audio track on
the first frame in release builds.

* Exit Demo in Release Builds when Music Stops

This change ensures that the demo correctly exits by itself when running
in release mode.

* Removed Unused Vulkan Device Features

Some of these unnecessary features were preventing the device from
initializing on Fiji.

* Add Read-Only Buffer Graph Binding Alias

This allows shaders to access scalar loads when appropriate.