Swapchain Dispatch + Example + Shader Binding Updates (#36)
* Swapchain Dispatch Support

This change adds support for direct dispatches that are dynamically
sized to cover the entire swapchain surface.

* Added Example Scene

This change adds a simple example scene to the example folder to
simplify testing.

* Shader Bindings Update & Y Direction Flip

This change adds inverse matrix varaints for all existing matrices
exposed to shaders. It also flips the Y direction of the projection
matrix to account for Vulkan's coordinate space.